# Conflicts:
#	conf/cassandra-env.sh
#	src/java/org/apache/cassandra/config/Config.java
#	src/java/org/apache/cassandra/config/DatabaseDescriptor.java
#	src/java/org/apache/cassandra/db/ClusteringPrefix.java
#	src/java/org/apache/cassandra/index/internal/CassandraIndex.java
#	src/java/org/apache/cassandra/io/sstable/ReducingKeyIterator.java
#	src/java/org/apache/cassandra/io/sstable/format/SSTableReader.java
#	src/java/org/apache/cassandra/io/sstable/format/SSTableScanner.java
#	src/java/org/apache/cassandra/io/sstable/format/big/BigTableReader.java
#	src/java/org/apache/cassandra/net/Message.java
#	src/java/org/apache/cassandra/net/MessagingService.java
#	src/java/org/apache/cassandra/schema/MigrationManager.java
#	src/java/org/apache/cassandra/schema/SchemaKeyspace.java
#	src/java/org/apache/cassandra/streaming/StreamReceiveTask.java
#	test/microbench/org/apache/cassandra/test/microbench/ReadWriteTestSmall.java
#	test/unit/org/apache/cassandra/cql3/KeyCacheCqlTest.java
#	test/unit/org/apache/cassandra/index/sasi/SASIIndexTest.java
#	test/unit/org/apache/cassandra/io/sstable/IndexSummaryManagerTest.java
#	test/unit/org/apache/cassandra/io/sstable/SSTableReaderTest.java
#	test/unit/org/apache/cassandra/io/sstable/SSTableWriterTest.java
#	test/unit/org/apache/cassandra/service/RemoveTest.java
#	test/unit/org/apache/cassandra/streaming/StreamingTransferTest.java

+++ test/unit/org/apache/cassandra/dht/KeyCollisionTest.java
@@ -27,6 +27,7 @@ import org.junit.Test;
 import org.apache.cassandra.SchemaLoader;
 import org.apache.cassandra.Util;
 import org.apache.cassandra.config.DatabaseDescriptor;
+import org.apache.cassandra.db.marshal.LongType;
 import org.apache.cassandra.schema.Schema;
 import org.apache.cassandra.db.ColumnFamilyStore;
 import org.apache.cassandra.db.Keyspace;
@@ -111,7 +112,7 @@ public class KeyCollisionTest
 
         public ByteSource asByteComparableSource()
         {
-            return IntegerType.instance.asByteComparableSource(IntegerType.instance.decompose(token));
+            return IntegerType.instance.asByteComparableSource(LongType.instance.decompose(token));
         }
 
         @Override
